+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### New features
 - Tint secret doors grey for the GM to differentiate them from regular doors
 - Toggle doors between secret and normal with ctrl+click
+- Makes the size of door controls independent of the scene's grid size
 
 ## v1.0.1
 - When adding a door to a synchronization group adjust it's state to bring it in sync with the other doors